+++ foxes/models/ground_models/no_ground.py
@@ -1,0 +1,10 @@
+from foxes.core import GroundModel
+
+class NoGround(GroundModel):
+    """
+    No ground effects
+    
+    :group: models.ground_models
+
+    """
+    pass