+++ foxes/models/partial_wakes/rotor_points.py
@@ -9,7 +9,6 @@ class RotorPoints(PartialWakesModel):
     :group: models.partial_wakes
 
     """
-
     def get_wake_points(self, algo, mdata, fdata):
         """
         Get the wake calculation points, and their
@@ -39,7 +38,7 @@ class RotorPoints(PartialWakesModel):
             rotor.from_data_or_store(rotor.RWEIGHTS, algo, mdata),
         )
 
-    def finalize_wakes_farm_calc(
+    def finalize_wakes(
         self,
         algo,
         mdata,